Move PeerConnectionEvent into separate module
