+++ crates/native/src/api/peer_connection_event/peer_connection_state.rs
@@ -1,0 +1,54 @@
+//! Current state of a [`PeerConnection`].
+
+use libwebrtc_sys as sys;
+
+#[cfg(doc)]
+use crate::{PeerConnection, api::IceConnectionState};
+
+/// Indicator of the current state of a [`PeerConnection`].
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum PeerConnectionState {
+    /// At least one of the connection's ICE transports is in the new state,
+    /// and none of them are in one of the following states: `connecting`,
+    /// `checking`, `failed`, `disconnected`, or all of the connection's
+    /// transports are in the `closed` state.
+    New,
+
+    /// One or more of the ICE transports are currently in the process of
+    /// establishing a connection. That is, their [`IceConnectionState`] is
+    /// either [`IceConnectionState::Checking`] or
+    /// [`IceConnectionState::Connected`], and no transports are in the
+    /// `failed` state.
+    Connecting,
+
+    /// Every ICE transport used by the connection is either in use (state
+    /// `connected` or `completed`) or is closed (state `closed`). In addition,
+    /// at least one transport is either `connected` or `completed`.
+    Connected,
+
+    /// At least one of the ICE transports for the connection is in the
+    /// `disconnected` state and none of the other transports are in the state
+    /// `failed`, `connecting` or `checking`.
+    Disconnected,
+
+    /// One or more of the ICE transports on the connection is in the `failed`
+    /// state.
+    Failed,
+
+    /// Peer connection is closed.
+    Closed,
+}
+
+impl From<sys::PeerConnectionState> for PeerConnectionState {
+    fn from(state: sys::PeerConnectionState) -> Self {
+        match state {
+            sys::PeerConnectionState::kNew => Self::New,
+            sys::PeerConnectionState::kConnecting => Self::Connecting,
+            sys::PeerConnectionState::kConnected => Self::Connected,
+            sys::PeerConnectionState::kDisconnected => Self::Disconnected,
+            sys::PeerConnectionState::kFailed => Self::Failed,
+            sys::PeerConnectionState::kClosed => Self::Closed,
+            _ => unreachable!(),
+        }
+    }
+}
