+++ lib/src/api/bridge/frb.dart
@@ -9,6 +9,12 @@ export 'api/media_stream_track.dart';
 export 'api/media_stream_track/media_type.dart';
 export 'api/media_stream_track/track_event.dart';
 export 'api/media_stream_track/media_track_state.dart';
+export 'api/peer_connection_event.dart';
+export 'api/peer_connection_event/ice_connection_state.dart';
+export 'api/peer_connection_event/ice_gathering_state.dart';
+export 'api/peer_connection_event/peer_connection_state.dart';
+export 'api/peer_connection_event/rtc_track_event.dart';
+export 'api/peer_connection_event/signaling_state.dart';
 export 'api/rtc_ice_candidate_stats.dart';
 export 'api/rtc_rtp_encoding_parameters.dart';
 export 'api/rtc_rtp_send_parameters.dart';
